flake: add more config
This fixes some things that @Gabriel439 suggested I fix, thanks!

Also enables creating machines that are either waifud hosts or waifud
runners. The distinction between the two is vauge and fuzzy at best at
the moment, but basically the waifud host will host the daemon and the
runners will be the machines actually running waifud.

The sshkeys in xeserv.waifud.runner.sshKeys are for waifud's agent.
Basically something is going to have to load the SSH keys into the
waifud-ssh-agent, but right now this is a "draw the rest of the owl"
kind of scenario. This will likely involve secrets of some kind, but I
will figure out the details for that later.

The configuration profiles will also allow waifud to do things properly
with zfs permission delegation and limited passwordless sudo. polkit
config is thrown into the loop so that waifud can manage libvirt without
having to SSH into target machines as root.

This should go a long way towards making it easier to run waifud on
NixOS machines. Documentation and examples to follow.

Signed-off-by: Xe <me@christine.website>